:sparkles: : view형식이 아닌 각각의 screen으로 수정 - 각 화면별로 네비게이팅 하기 위해 수정
